backend-server/requirements.txt
accelerate==0.27.2
aiohttp==3.9.3
aiohttp-retry==2.8.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.7.2
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
azure-cognitiveservices-speech==1.36.0
bcrypt==4.1.2
beartype==0.17.2
bidict==0.23.1
bitsandbytes==0.42.0
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
CoLT5-attention==0.10.19
cryptography==42.0.5
dataclasses-json==0.6.4
datasets==2.18.0
decorator==5.1.1
Deprecated==1.2.14
dill==0.3.8
distro==1.9.0
dnspython==2.6.1
einops==0.7.0
einops-exts==0.0.4
einx==0.1.3
eventlet==0.35.2
exceptiongroup==1.2.0
executing==2.0.1
fairscale==0.4.13
filelock==3.13.1
Flask==3.0.2
Flask-Cors==4.0.0
Flask-SocketIO==5.3.6
frozendict==2.4.0
frozenlist==1.4.1
fsspec==2024.2.0
future==0.18.3
google-ai-generativelanguage==0.4.0
google-api-core==2.17.1
google-auth==2.28.1
google-generativeai==0.4.0
googleapis-common-protos==1.62.0
gpt3-tokenizer==0.1.4
greenlet==3.0.3
grpcio==1.62.0
grpcio-status==1.62.0
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
huggingface-hub==0.21.4
idna==3.6
importlib-metadata==7.0.1
iniconfig==2.0.0
install==1.3.5
ipython==8.18.1
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
jsonpatch==1.33
jsonpointer==2.4
langchain==0.1.11
langchain-community==0.0.25
langchain-core==0.1.29
langchain-text-splitters==0.0.1
langsmith==0.1.19
lion-pytorch==0.1.2
local-attention==1.9.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
mypy-extensions==1.0.0
networkx==3.2.1
nodeenv==1.8.0
numpy==1.26.4
openai==1.13.3
orjson==3.9.15
packaging==23.2
pandas==2.2.1
parso==0.8.3
pexpect==4.9.0
pillow==10.2.0
pluggy==1.4.0
prisma==0.12.0
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.3
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==15.0.1
pyarrow-hotfix==0.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.6.3
pydantic_core==2.16.3
Pygments==2.17.2
PyJWT==2.8.0
pytest==8.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.9.0
python-socketio==5.11.1
pytz==2024.1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
rich==13.7.1
ring-attention-pytorch==0.2.7
rsa==4.9
safetensors==0.4.2
scipy==1.12.0
sentencepiece==0.2.0
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.28
stack-data==0.6.3
StrEnum==0.4.15
sympy==1.12
tenacity==8.2.3
tiktoken==0.6.0
timm==0.9.16
tokenizers==0.15.2
tokenmonster==1.1.12
tomli==2.0.1
tomlkit==0.12.4
torch==2.2.1
torchdiffeq==0.2.3
torchvision==0.17.1
tqdm==4.66.2
traitlets==5.14.1
transformers==4.38.2
twilio==9.0.0
typing==3.7.4.3
typing-inspect==0.9.0
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
vector-quantize-pytorch==1.12.0
vonage==3.13.0
vonage-jwt==1.1.0
wcwidth==0.2.13
websockets==12.0
Werkzeug==3.0.1
wrapt==1.16.0
wsproto==1.2.0
xxhash==3.4.1
yarl==1.9.4
zetascale==0.9.1
zipp==3.17.0
